Move and refactor
