+++ api_tests/ledger_query_service/test.js
@@ -1,15 +1,15 @@
 const chai = require("chai");
 chai.use(require("chai-http"));
 const should = chai.should();
-const bitcoin_rpc_client_conf = require("../bitcoin_rpc_client_conf.js");
-const web3_conf = require("../web3_conf.js");
-const wallet_conf = require("../wallet_conf.js");
-const lqs_conf = require("../ledger_query_service_conf.js");
+const bitcoin = require("../lib/bitcoin.js");
+const web3_conf = require("../lib/web3_conf.js");
+const wallet = require("../lib/wallet.js");
+const lqs_conf = require("../lib/lqs.js");
 
-const bitcoin_rpc_client = bitcoin_rpc_client_conf.create_client();
+const bitcoin_rpc_client = bitcoin.create_client();
 const lqs = lqs_conf.create("localhost", 8080);
 const web3 = web3_conf.create();
-const wallet = wallet_conf.create();
+const toby_wallet = wallet.create("toby");
 
 function sleep(ms) {
     return new Promise(resolve => setTimeout(resolve, ms));
@@ -18,9 +18,9 @@ function sleep(ms) {
 describe("Test Ledger Query Service API", () => {
     before(async function() {
         this.timeout(5000);
-        await bitcoin_rpc_client_conf.btc_activate_segwit();
-        await wallet.btc().fund(5);
-        await wallet.eth().fund(20);
+        await bitcoin.btc_activate_segwit();
+        await toby_wallet.btc().fund(5);
+        await toby_wallet.eth().fund(20);
     });
 
     describe("Bitcoin", () => {
@@ -64,7 +64,7 @@ describe("Test Ledger Query Service API", () => {
 
             it("LQS should respond with transaction match when requesting on the `to_address` bitcoin transaction query", async function() {
                 this.slow(1000);
-                return wallet
+                return toby_wallet
                     .btc()
                     .send_btc_to_address(to_address, 100000000)
                     .then(() => {
@@ -182,7 +182,7 @@ describe("Test Ledger Query Service API", () => {
     describe("Ethereum", () => {
         describe("Transactions", () => {
             before(async () => {
-                await wallet.eth().fund(10);
+                await toby_wallet.eth().fund(10);
             });
 
             it("LQS should respond not found when getting a non-existent ethereum transaction query", async function() {
@@ -222,7 +222,7 @@ describe("Test Ledger Query Service API", () => {
             });
 
             it("LQS should respond with no transaction match (yet) when requesting on the `to_address` ethereum block query", async function() {
-                return wallet
+                return toby_wallet
                     .eth()
                     .send_eth_transaction_to(
                         "0xaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
@@ -245,7 +245,7 @@ describe("Test Ledger Query Service API", () => {
 
             it("LQS should respond with transaction match when requesting on the `to_address` ethereum transaction query", async function() {
                 this.slow(2000);
-                return wallet
+                return toby_wallet
                     .eth()
                     .send_eth_transaction_to(to_address, "", 5)
                     .then(() => {
@@ -315,7 +315,7 @@ describe("Test Ledger Query Service API", () => {
                 this.slow(6000);
                 return sleep(3000)
                     .then(() => {
-                        return wallet
+                        return toby_wallet
                             .eth()
                             .send_eth_transaction_to(
                                 "0xaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
