+++ api_tests/lib/bitcoin.js
@@ -1,0 +1,156 @@
+const bitcoin = require("bitcoinjs-lib");
+const BitcoinRpcClient = require("bitcoin-core");
+const web3_conf = require("./web3_conf.js");
+const web3 = web3_conf.create();
+
+const util = require("./util.js");
+
+//FIXME: Remove this whenever this change:
+// https://github.com/bitcoinjs/bitcoinjs-lib/commit/44a98c0fa6487eaf81500427366787a953ff890d#diff-9e60abeb4e2333a5d2f02de53b4edfac
+// Hits npm!
+const regtest = {
+    messagePrefix: "\x18Bitcoin Signed Message:\n",
+    bech32: "bcrt",
+    bip32: {
+        public: 0x043587cf,
+        private: 0x04358394,
+    },
+    pubKeyHash: 0x6f,
+    scriptHash: 0xc4,
+    wif: 0xef,
+};
+
+
+let _bitcoin_rpc_client;
+
+function create_bitcoin_rpc_client() {
+    const btc_config = global.harness.ledgers_config.bitcoin;
+    if (!btc_config) {
+        throw new Error("ledger.bitcoin configuration is needed");
+    }
+    return (_bitcoin_rpc_client =
+        _bitcoin_rpc_client ||
+        new BitcoinRpcClient({
+            network: "regtest",
+            port: btc_config.rpc_port,
+            host: btc_config.rpc_host,
+            username: btc_config.rpc_username,
+            password: btc_config.rpc_password,
+        }));
+}
+
+module.exports.create_client = () => {
+    return create_bitcoin_rpc_client();
+};
+
+module.exports.btc_generate = async function(num = 1) {
+    return create_bitcoin_rpc_client().generate(num);
+};
+
+module.exports.btc_activate_segwit = async function() {
+    return create_bitcoin_rpc_client().generate(432);
+};
+
+module.exports.btc_import_address = async function(address) {
+    return create_bitcoin_rpc_client().importAddress(address);
+};
+
+async function btc_balance(address) {
+    let btc_balance = await _bitcoin_rpc_client.getReceivedByAddress(address);
+    return parseFloat(btc_balance) * 100000000;
+}
+
+module.exports.btc_balance = async function(address) {
+    return btc_balance(address);
+};
+
+module.exports.log_btc_balance = async function(
+    when,
+    player,
+    address,
+    address_type
+) {
+    util
+        .logger()
+        .info(
+            "%s the swap, %s has %s satoshis at the %s address %s",
+            when,
+            player,
+            await btc_balance(address),
+            address_type,
+            address
+        );
+};
+
+module.exports.erc20_balance = async function(
+    token_holder_address,
+    contract_address
+) {
+    const function_identifier = "70a08231";
+
+    const padded_address = token_holder_address
+        .replace(/^0x/, "")
+        .padStart(64, "0");
+    const payload = "0x" + function_identifier + padded_address;
+
+    const tx = {
+        from: token_holder_address,
+        to: contract_address,
+        data: payload,
+    };
+
+    let hex_balance = await web3.eth.call(tx);
+    return web3.utils.toBN(hex_balance);
+};
+
+class BitcoinWallet {
+    constructor() {
+        this.keypair = bitcoin.ECPair.makeRandom({ rng: util.test_rng });
+        this.bitcoin_utxos = [];
+        this._identity = bitcoin.payments.p2wpkh({
+            pubkey: this.keypair.publicKey,
+            network: regtest,
+        });
+    }
+
+    identity() {
+        return this._identity;
+    }
+
+    async fund(btc_value) {
+        let txid = await _bitcoin_rpc_client.sendToAddress(
+            this.identity().address,
+            btc_value
+        );
+        let raw_transaction = await _bitcoin_rpc_client.getRawTransaction(txid);
+        let transaction = bitcoin.Transaction.fromHex(raw_transaction);
+        for (let [i, out] of transaction.outs.entries()) {
+            if (out.script.equals(this.identity().output)) {
+                out.txid = txid;
+                out.vout = i;
+                this.bitcoin_utxos.push(out);
+            }
+        }
+    }
+
+    async send_btc_to_address(to, value) {
+        const txb = new bitcoin.TransactionBuilder();
+        const utxo = this.bitcoin_utxos.shift();
+        const to_address = bitcoin.address.fromBech32(to);
+        const input_amount = utxo.value;
+        const key_pair = this.keypair;
+        const fee = 2500;
+        const change = input_amount - value - fee;
+        txb.addInput(utxo.txid, utxo.vout, null, this.identity().output);
+        //TODO: Add it back to UTXOs after transaction is successful
+        txb.addOutput(this.identity().output, change);
+        txb.addOutput(bitcoin.address.toOutputScript(to, regtest), value);
+        txb.sign(0, key_pair, null, null, input_amount);
+
+        return _bitcoin_rpc_client.sendRawTransaction(txb.build().toHex());
+    }
+}
+
+module.exports.create_wallet = () => {
+    return new BitcoinWallet();
+};
